updating the cube route to show a different way of doing it
