+++ routes/cube/index.js
@@ -1,0 +1,6 @@
+const router = require('express').Router();
+
+router.use('/api', require('./api'));
+router.use('', require('./views'));
+
+module.exports = router;
